Show Favorite chains in NetworkSelector Button (#2644)
<!-- start pr-codex -->

## PR-Codex overview
The focus of this PR is to refactor the favorite chains feature, adding a new hook `useFavoriteChains` and integrating it into various components.

### Detailed summary
- Added `bundledDependencies: false` in `pnpm-lock.yaml`
- Renamed `useFavouriteChains` to `useFavouriteChainIds`
- Introduced `useFavoriteChains` hook for fetching and handling favorite chains
- Updated components to use `useFavoriteChains` for favorite chains data
- Reorganized network selector button to display favorite chains
- Refactored `CustomConnectWallet` to utilize `useFavoriteChains` for favorite chains display

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->
